refactor: 상수파일을 메세지 상수와 그외로 분리 및 소문자로 수정
