+++ src/component/common/CardCoButton.tsx
@@ -1,4 +1,4 @@
-import { CARD_CO_NAME } from "../../CONSTANT";
+import { CARD_CO_NAME } from "../../constant/message";
 import { CardCo } from "../../type";
 import { GetCardCoIcon } from "../../util/getCardCoIcon";
