+++ src/validation/cardOwner.ts
@@ -1,4 +1,4 @@
-import { REGULAR_EXPRESSION } from "../CONSTANT";
+import { REGULAR_EXPRESSION } from "../constant/etc";
 
 export const validateCardOwner = (owner: string) => {
   const regExp = REGULAR_EXPRESSION.CAPITAL_LETTER_OR_SPACE;
