+++ src/validation/number.ts
@@ -1,4 +1,4 @@
-import { REGULAR_EXPRESSION } from "../CONSTANT";
+import { REGULAR_EXPRESSION } from "../constant/etc";
 
 export const validateNumber = (value: string) => {
   const regExp = REGULAR_EXPRESSION.ONE_NUMBER;
